Support NZTM mbtiles creation.
